+++ bundles/ng2-rike.umd.js
@@ -1422,11 +1422,6 @@ var RikeTargetImpl = (function (_super) {
                     this._observer.error(cancel);
                     this._rikeEvents.emit(cancel);
                 }
-                catch (e) {
-                    console.error("Failed to cancel Rike operation", e);
-                    this._rikeEvents.error(new RikeExceptionEvent(this._operation.operation, e));
-                    throw e;
-                }
                 finally {
                     this._operation = undefined;
                     try {
@@ -1472,18 +1467,9 @@ var RikeTargetImpl = (function (_super) {
                 }
             };
             _this._subscr = response.subscribe(function (httpResponse) {
-                try {
-                    var response_1 = operation.protocol.readResponse(httpResponse);
-                    responseObserver.next(response_1);
-                    _this._rikeEvents.emit(new RikeSuccessEvent(operation, response_1));
-                }
-                catch (e) {
-                    console.error("Failed to handle Rike response", e);
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e, {
-                        response: httpResponse,
-                        error: e
-                    }));
-                }
+                var response = operation.protocol.readResponse(httpResponse);
+                responseObserver.next(response);
+                _this._rikeEvents.emit(new RikeSuccessEvent(operation, response));
             }, function (error) {
                 console.error("[" + _this.target + "] " + operation.name + " failed", error);
                 var errorResponse = toErrorResponse(error);
@@ -1492,21 +1478,12 @@ var RikeTargetImpl = (function (_super) {
                     responseObserver.error(errorResponse);
                     _this._rikeEvents.emit(new RikeErrorResponseEvent(operation, errorResponse));
                 }
-                catch (e) {
-                    console.error("Failed to handle Rike error", e);
-                    errorResponse.error = e;
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e, errorResponse));
-                }
                 finally {
                     cleanup();
                 }
             }, function () {
                 try {
                     responseObserver.complete();
-                }
-                catch (e) {
-                    console.error("Failed to complete Rike response", e);
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e));
                 }
                 finally {
                     cleanup();
@@ -1583,7 +1560,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.request(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1594,7 +1571,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.request(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1605,7 +1582,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.get(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1616,7 +1593,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.post(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1627,7 +1604,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.put(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1639,7 +1616,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.delete(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1650,7 +1627,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.patch(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1661,7 +1638,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.head(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -1857,7 +1834,7 @@ var RikeStatusComponent = (function () {
     ], RikeStatusComponent.prototype, "rikeStatusLabelClass", null);
     RikeStatusComponent = __decorate$3([
         _angular_core.Component({
-            selector: '[rikeStatus],[rikeStatusLabels],[rikeStatusLabelText],[rikeStatusLabelClass]',
+            selector: 'rike-status,[rikeStatus],[rikeStatusLabels],[rikeStatusLabelText],[rikeStatusLabelClass]',
             template: "<span class=\"rike-status-icon\"></span> {{text}}",
             host: {
                 "[class]": "cssClass",
@@ -2428,7 +2405,7 @@ var RikeErrorsComponent = (function () {
     ], RikeErrorsComponent.prototype, "rikeErrorsOf", null);
     RikeErrorsComponent = __decorate$4([
         _angular_core.Component({
-            selector: '[rikeErrors],[rikeErrorsOf]',
+            selector: 'rike-errors,[rikeErrors],[rikeErrorsOf]',
             template: "\n    <ul class=\"rike-error-list\" *ngIf=\"errors.length\">\n        <li class=\"rike-error\" *ngFor=\"let error of errors\">{{error.message}}</li>\n    </ul>\n    ",
             host: {
                 "[class.rike-errors]": "true",
@@ -2468,6 +2445,138 @@ function provideEventSource(_a) {
         },
     ];
 }
+
+var __decorate$6 = (undefined && undefined.__decorate) || function (decorators, target, key, desc) {
+    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
+    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
+    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
+    return c > 3 && r && Object.defineProperty(target, key, r), r;
+};
+var __metadata$6 = (undefined && undefined.__metadata) || function (k, v) {
+    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+};
+var __param$4 = (undefined && undefined.__param) || function (paramIndex, decorator) {
+    return function (target, key) { decorator(target, key, paramIndex); }
+};
+var RikeDisabledDirective = (function () {
+    function RikeDisabledDirective(_rikeDisabledBy) {
+        this._rikeDisabledBy = _rikeDisabledBy;
+        this._disabledByDefault = false;
+    }
+    Object.defineProperty(RikeDisabledDirective.prototype, "disabled", {
+        get: function () {
+            return this.rikeDisabledBy.processing || this._disabledByDefault;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    Object.defineProperty(RikeDisabledDirective.prototype, "rikeDisabled", {
+        set: function (disabled) {
+            this._disabledByDefault = !!disabled;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    Object.defineProperty(RikeDisabledDirective.prototype, "rikeDisabledBy", {
+        get: function () {
+            return this._rikeDisabledBy;
+        },
+        set: function (collector) {
+            this._rikeDisabledBy = collector;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    __decorate$6([
+        _angular_core.HostBinding("disabled"),
+        _angular_core.HostBinding("class.rike-disabled"), 
+        __metadata$6('design:type', Boolean)
+    ], RikeDisabledDirective.prototype, "disabled", null);
+    __decorate$6([
+        _angular_core.Input(), 
+        __metadata$6('design:type', Object), 
+        __metadata$6('design:paramtypes', [Object])
+    ], RikeDisabledDirective.prototype, "rikeDisabled", null);
+    __decorate$6([
+        _angular_core.Input(), 
+        __metadata$6('design:type', StatusCollector)
+    ], RikeDisabledDirective.prototype, "rikeDisabledBy", null);
+    RikeDisabledDirective = __decorate$6([
+        _angular_core.Directive({
+            selector: '[rikeDisabled],[rikeDisabledBy]',
+            exportAs: 'rikeDisabled',
+        }),
+        __param$4(0, _angular_core.Optional()), 
+        __metadata$6('design:paramtypes', [StatusCollector])
+    ], RikeDisabledDirective);
+    return RikeDisabledDirective;
+}());
+
+var __decorate$7 = (undefined && undefined.__decorate) || function (decorators, target, key, desc) {
+    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
+    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
+    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
+    return c > 3 && r && Object.defineProperty(target, key, r), r;
+};
+var __metadata$7 = (undefined && undefined.__metadata) || function (k, v) {
+    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+};
+var __param$5 = (undefined && undefined.__param) || function (paramIndex, decorator) {
+    return function (target, key) { decorator(target, key, paramIndex); }
+};
+var RikeReadonlyDirective = (function () {
+    function RikeReadonlyDirective(_rikeReadonlyBy) {
+        this._rikeReadonlyBy = _rikeReadonlyBy;
+        this._readonlyByDefault = false;
+    }
+    Object.defineProperty(RikeReadonlyDirective.prototype, "readonly", {
+        get: function () {
+            return this.rikeReadonlyBy.processing || this._readonlyByDefault;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    Object.defineProperty(RikeReadonlyDirective.prototype, "rikeReadonly", {
+        set: function (disabled) {
+            this._readonlyByDefault = !!disabled;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    Object.defineProperty(RikeReadonlyDirective.prototype, "rikeReadonlyBy", {
+        get: function () {
+            return this._rikeReadonlyBy;
+        },
+        set: function (collector) {
+            this._rikeReadonlyBy = collector;
+        },
+        enumerable: true,
+        configurable: true
+    });
+    __decorate$7([
+        _angular_core.HostBinding("readonly"),
+        _angular_core.HostBinding("class.rike-readonly"), 
+        __metadata$7('design:type', Boolean)
+    ], RikeReadonlyDirective.prototype, "readonly", null);
+    __decorate$7([
+        _angular_core.Input(), 
+        __metadata$7('design:type', Object), 
+        __metadata$7('design:paramtypes', [Object])
+    ], RikeReadonlyDirective.prototype, "rikeReadonly", null);
+    __decorate$7([
+        _angular_core.Input(), 
+        __metadata$7('design:type', StatusCollector)
+    ], RikeReadonlyDirective.prototype, "rikeReadonlyBy", null);
+    RikeReadonlyDirective = __decorate$7([
+        _angular_core.Directive({
+            selector: '[rikeReadonly],[rikeReadonlyBy]',
+            exportAs: 'rikeReadonly',
+        }),
+        __param$5(0, _angular_core.Optional()), 
+        __metadata$7('design:paramtypes', [StatusCollector])
+    ], RikeReadonlyDirective);
+    return RikeReadonlyDirective;
+}());
 
 var __extends$4 = (undefined && undefined.__extends) || function (d, b) {
     for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
@@ -2898,10 +3007,14 @@ var RikeModule = (function () {
             declarations: [
                 RikeStatusComponent,
                 RikeErrorsComponent,
+                RikeDisabledDirective,
+                RikeReadonlyDirective,
             ],
             exports: [
                 RikeStatusComponent,
                 RikeErrorsComponent,
+                RikeDisabledDirective,
+                RikeReadonlyDirective,
             ],
         }), 
         __metadata('design:paramtypes', [])
@@ -2910,6 +3023,7 @@ var RikeModule = (function () {
 }());
 
 exports.RikeModule = RikeModule;
+exports.RikeDisabledDirective = RikeDisabledDirective;
 exports.ErrorCollector = ErrorCollector;
 exports.RikeErrorsComponent = RikeErrorsComponent;
 exports.RikeEventSource = RikeEventSource;
@@ -2930,6 +3044,7 @@ exports.Protocol = Protocol;
 exports.JSON_PROTOCOL = JSON_PROTOCOL;
 exports.jsonProtocol = jsonProtocol;
 exports.HTTP_PROTOCOL = HTTP_PROTOCOL;
+exports.RikeReadonlyDirective = RikeReadonlyDirective;
 exports.Resource = Resource;
 exports.RikeResource = RikeResource;
 exports.LoadableResource = LoadableResource;