+++ ng2-rike.d.ts
@@ -1,4 +1,5 @@
 import { RikeOptionsArgs } from "./ng2-rike/options";
+export * from "./ng2-rike/disabled.directive";
 export * from "./ng2-rike/error-collector";
 export * from "./ng2-rike/errors.component";
 export * from "./ng2-rike/event";
@@ -6,6 +7,7 @@ export * from "./ng2-rike/event-source-provider";
 export * from "./ng2-rike/field-error";
 export * from "./ng2-rike/options";
 export * from "./ng2-rike/protocol";
+export * from "./ng2-rike/readonly.directive";
 export * from "./ng2-rike/resource";
 export * from "./ng2-rike/resource-provider";
 export * from "./ng2-rike/rike";