+++ ng2-rike/rike.js
@@ -18,9 +18,10 @@ var __param = (this && this.__param) || function (paramIndex, decorator) {
 import { Injectable, Optional, EventEmitter } from "@angular/core";
 import { Response, Http, RequestMethod, RequestOptions, ResponseOptions, ResponseType } from "@angular/http";
 import { Observable } from "rxjs/Rx";
-import { RikeSuccessEvent, RikeOperationEvent, RikeCancelEvent, RikeErrorResponseEvent, RikeExceptionEvent } from "./event";
+import { RikeSuccessEvent, RikeOperationEvent, RikeCancelEvent, RikeErrorResponseEvent } from "./event";
 import { RikeOptions, DEFAULT_RIKE_OPTIONS, relativeUrl } from "./options";
 import { HTTP_PROTOCOL, jsonProtocol } from "./protocol";
+import { RikeExceptionEvent } from "./event";
 var REQUEST_METHODS = {
     "GET": RequestMethod.Get,
     "POST": RequestMethod.Post,
@@ -385,11 +386,6 @@ var RikeTargetImpl = (function (_super) {
                     this._observer.error(cancel);
                     this._rikeEvents.emit(cancel);
                 }
-                catch (e) {
-                    console.error("Failed to cancel Rike operation", e);
-                    this._rikeEvents.error(new RikeExceptionEvent(this._operation.operation, e));
-                    throw e;
-                }
                 finally {
                     this._operation = undefined;
                     try {
@@ -435,18 +431,9 @@ var RikeTargetImpl = (function (_super) {
                 }
             };
             _this._subscr = response.subscribe(function (httpResponse) {
-                try {
-                    var response_1 = operation.protocol.readResponse(httpResponse);
-                    responseObserver.next(response_1);
-                    _this._rikeEvents.emit(new RikeSuccessEvent(operation, response_1));
-                }
-                catch (e) {
-                    console.error("Failed to handle Rike response", e);
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e, {
-                        response: httpResponse,
-                        error: e
-                    }));
-                }
+                var response = operation.protocol.readResponse(httpResponse);
+                responseObserver.next(response);
+                _this._rikeEvents.emit(new RikeSuccessEvent(operation, response));
             }, function (error) {
                 console.error("[" + _this.target + "] " + operation.name + " failed", error);
                 var errorResponse = toErrorResponse(error);
@@ -455,21 +442,12 @@ var RikeTargetImpl = (function (_super) {
                     responseObserver.error(errorResponse);
                     _this._rikeEvents.emit(new RikeErrorResponseEvent(operation, errorResponse));
                 }
-                catch (e) {
-                    console.error("Failed to handle Rike error", e);
-                    errorResponse.error = e;
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e, errorResponse));
-                }
                 finally {
                     cleanup();
                 }
             }, function () {
                 try {
                     responseObserver.complete();
-                }
-                catch (e) {
-                    console.error("Failed to complete Rike response", e);
-                    _this._rikeEvents.error(new RikeExceptionEvent(operation, e));
                 }
                 finally {
                     cleanup();
@@ -546,7 +524,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.request(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -557,7 +535,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.request(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -568,7 +546,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.get(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -579,7 +557,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.post(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -590,7 +568,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.put(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -602,7 +580,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.delete(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -613,7 +591,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.patch(this.requestUrl(options), options.body, options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };
@@ -624,7 +602,7 @@ var RikeOperationImpl = (function (_super) {
             return this.wrapResponse(this.internals.head(this.requestUrl(options), options));
         }
         catch (e) {
-            this.target.rikeEvents.error(new RikeExceptionEvent(this, e));
+            this.target.rikeEvents.emit(new RikeExceptionEvent(this, e));
             throw e;
         }
     };