+++ CHANGELOG.md
@@ -1,8 +1,23 @@
+# 0.2.3 (2016-10-04)
+
+### Features
+
+* Add `RikeDisabledDirective` to disable controls while Rike operation is in process.
+* Add `RokeReadonlyDirective` to mark input controls red only while Rike operation is in process.
+* `RikeErrorsComponent` additionally bound to `rike-errors` element.
+* `RikeStatusComponent` additionally bound to `rike-status` element.
+
+### Bug Fixes
+
+* Never send errors via Rike events emitter. Emit events instead.
+
+
 # 0.2.2 (2016-09-25)
 
 ### Peer Dependencies
 
 * Depend on Angular 2.0.0. or higher.
+
 
 # O.2.1 (2016-09-21)
 
@@ -11,6 +26,7 @@
 * *`provideResource`:* Permit multiple anonymous (without `provide` key) resources at the same level.
 * Fix reporting of operation cancellation.
 
+
 # 0.2.0 (2016-09-15)
 
 ### Peer Dependencies
@@ -18,11 +34,13 @@
 * Angular 2 release.
 * Peer dependencies are also dev dependencies now.
 
+
 # 0.1.5 (2016-09-13)
 
 ### Bugs
 
 * *`ErrorCollector`:* Display readable message when operation cancelled.
+
 
 # 0.1.4 (2016-09-11)
 
@@ -30,6 +48,7 @@
 
 * *`Protocol`:* More protocol modification methods.
 * *`CRUDResource`:* More customizations.
+
 
 # 0.1.3 (2016-09-08)
 
@@ -42,6 +61,7 @@
 * *npm package:* Use peer dependencies.
 * *specs:* Use WebPack to bundle specs.
 
+
 # 0.1.2 (2016-09-03)
 
 ### Features
@@ -52,6 +72,7 @@
 ### Peer Dependencies
 
 * Angular2 rc6.
+
 
 # 0.1.1 (2016-09-02)